perf(vggt/jax): jit aggregator with padded KV-cache (WIP, integration drift)
Replaces the growing-list KV cache with a fixed-size padded cache so the
aggregator can be traced once and reused across streaming frames. Makes
the jit static-arg key only distinguish first-frame vs subsequent-frame
(bool) instead of per-frame index, so we compile twice instead of once
per frame.

Measured (profile_streaming, H100):
- aggregator: 2897ms -> 57ms (51x)
- total extract: 4040ms -> 1255ms (3.2x)
- camera_head still eager, now ~95% of remaining time

Known issue: integration test test_rollout_parity fails with
world_points drift ~0.36 vs 1e-2 tolerance. Fast parity (8/8) and
Level-3 cache parity (12/12) pass, but those only exercise the
legacy 2-tuple path via past_kvs=None. The padded 3-tuple path
used by JAXVGGTFeatureExtractor is not yet covered by parity
and is the suspected source of the drift (jax.lax.cond eviction
plumbing or _padded_evict zero-fill semantics).

Integration test fixture pinned to fp32 to rule out precision;
drift is identical at bf16 and fp32, confirming it is a cache-state
bug rather than a kernel-precision issue.

Adds profile_streaming.py for per-phase latency attribution.

+++ modules/vggt/tests/test_jax_integration.py
@@ -110,9 +110,13 @@ class TestJAXvsPyTorchExtractor:
 
     @pytest.fixture(scope="class")
     def jax_ext(self):
+        import jax.numpy as jnp
+
         from modules.vggt.jax import JAXVGGTFeatureExtractor
 
-        return JAXVGGTFeatureExtractor(device="cuda")
+        # Force fp32 to line up with the fp32-pinned PyTorch fixture below;
+        # default bf16 would exceed ROLLOUT_ATOL against fp32 PyTorch.
+        return JAXVGGTFeatureExtractor(device="cuda", dtype=jnp.float32)
 
     @pytest.fixture(scope="class")
     def pt_ext(self):
